# openapi.yaml
openapi: 3.1.0
info:
  title: Marconi Inmobiliaria Analytics & Geocoding API
  description: |
    Comprehensive API for real estate analytics tracking and geocoding services.
    Includes GDPR-compliant user session management, property view tracking,
    lead generation attribution, and external geocoding proxy services.
  version: "1.0.0"
  contact:
    name: Marconi Inmobiliaria API Support
    email: contact@marconi-inmobiliaria.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://marconi-inmobiliaria.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    StandardResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Operation success status
        error:
          type: string
          description: Error message if success is false
        message:
          type: string
          description: Success message

    # ==== GEOCODING SCHEMAS ====
    GeocodingResponse:
      type: object
      properties:
        lat:
          type: number
          format: float
          description: Latitude coordinate
          example: -29.15
        lng:
          type: number
          format: float
          description: Longitude coordinate
          example: -59.65
        display_name:
          type: string
          description: Human-readable address description
          example: "Reconquista, Santa Fe, Argentina"

    # ==== ANALYTICS SESSION SCHEMAS ====
    CreateSessionRequest:
      type: object
      properties:
        user_agent:
          type: string
          description: Browser user agent string
          maxLength: 500
        country_code:
          type: string
          description: ISO 3166-1 alpha-2 country code
          pattern: '^[A-Z]{2}$'
          example: "AR"
        device_type:
          type: string
          enum: [desktop, mobile, tablet]
          description: Device type classification
        browser:
          type: string
          description: Browser name
          maxLength: 100
        os:
          type: string
          description: Operating system name
          maxLength: 100
        referrer_domain:
          type: string
          description: Referring domain (without protocol)
          maxLength: 255
        landing_page:
          type: string
          description: First page visited in session
          maxLength: 500
        utm_source:
          type: string
          description: UTM source parameter
          maxLength: 255
        utm_medium:
          type: string
          description: UTM medium parameter
          maxLength: 255
        utm_campaign:
          type: string
          description: UTM campaign parameter
          maxLength: 255
        utm_term:
          type: string
          description: UTM term parameter
          maxLength: 255
        utm_content:
          type: string
          description: UTM content parameter
          maxLength: 255
    SessionResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            session_id:
              type: string
              format: uuid
              description: Unique session identifier

    # ==== PROPERTY VIEW SCHEMAS ====
...
    RateLimitErrorResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            retry_after:
              type: integer
              description: Seconds to wait before retrying
paths:
...
  /geocode:
    get:
      tags:
        - Geocoding
      summary: Geocode an address to coordinates
      description: |
        Converts an address to latitude/longitude coordinates using OpenStreetMap Nominatim.
        Features 24-hour caching, rate limiting (1 request/second), and Argentina-specific filtering.
        Provides fallback coordinates for Reconquista, Santa Fe on errors or no results.
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 500
          description: Address to geocode
          example: "San Martin 123, Reconquista, Santa Fe"
      responses:
        '200':
          description: Geocoding successful (includes fallback responses)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeocodingResponse'
              examples:
                success:
                  summary: Successful geocoding
                  value:
                    lat: -29.15
                    lng: -59.65
                    display_name: "San Martin 123, Reconquista, Santa Fe, Argentina"
                fallback:
                  summary: Fallback response
                  value:
                    lat: -29.15
                    lng: -59.65
                    display_name: "Reconquista, Santa Fe, Argentina (fallback)"
        '400':
          description: Invalid request - missing or empty address parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                success: false
                error: "Address parameter is required"
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                example: 1
              description: Seconds to wait before retrying
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitErrorResponse'

  # ==== ANALYTICS SESSION ENDPOINTS ====
  /analytics/session:
    post:
      tags:
        - Analytics
      summary: Create or retrieve analytics session
      description: |
        Creates a new analytics session or retrieves existing one based on IP and user agent.
        IP addresses are hashed for GDPR compliance. Sessions are deduplicated within 4-hour windows.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Session created or retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
              example:
                success: true
                session_id: "550e8400-e29b-41d4-a716-446655440000"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'

    put:
      tags:
        - Analytics
      summary: Update session last seen timestamp
      description: Updates the last_seen_at timestamp for an existing session
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID to update
      responses:
        '200':
          description: Session updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '400':
          description: Missing or invalid session ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'

  # ==== PROPERTY VIEW ENDPOINTS ====
...